Add files via upload
Instructions
Use prior sections of this unit in which you are shown how to get input from the user via 1) prompt, 2) input field. You will be creating a .html file, and adding JavaScript in between script tags. The idea is to get the user to type in input and echo the input back.

You will

1. create a .html file, with basic tags for HTML5

2. add a label of 'What language was created to add interactivity to the Web?  :', and a corresponding input field.

3. add <script>...</script> tags

4. add JavaScript code that lifts what the user typed into the input field (in step 2 above) and displays it with an alert box

5. add code that issues a prompt and asks the user what code library they would prefer - give choices of Vue, React or jQuery.

6. show the choice the user typed into the prompt dialog, with an alert box